move down js_crypto
