--- src/js_crypto/dune
@@ -1,7 +1,0 @@
-(rule
- (targets constants.ts)
- (mode promote)
- (action
-  (with-stdout-to
-   constants.ts
-   (run %{exe:../snarkyjs-bindings/ocaml/snarky_js_constants.exe}))))
