--- src/js_crypto/random.ts
@@ -1,6 +1,0 @@
-import { randomBytes as randomBytesNode } from 'crypto';
-export { randomBytes };
-
-function randomBytes(n: number) {
-  return new Uint8Array(randomBytesNode(n));
-}
