--- src/js_crypto/random.web.ts
@@ -1,5 +1,0 @@
-export { randomBytes };
-
-function randomBytes(n: number) {
-  return crypto.getRandomValues(new Uint8Array(n));
-}
