--- src/js_crypto/test_vectors/dune
@@ -1,12 +1,0 @@
-(rule
- (targets poseidonKimchi.ts poseidonLegacy.ts)
- (mode promote)
- (deps
-  (:script ./dump-test-vectors.sh)
-  (source_tree ../../../../crypto/proof-systems)
-  ../../../.prettierrc.cjs
-  (glob_files ../../../node_modules/prettier/*))
- (action
-  (progn
-   (run chmod -R +w ../../../../crypto/proof-systems/ .)
-   (run %{script} poseidonKimchi.ts poseidonLegacy.ts))))
