--- src/js_crypto/test_vectors/poseidonKimchi.ts
@@ -1,57 +1,0 @@
-// @gen this file is generated - don't edit it directly
-export { testPoseidonKimchiFp };
-let testPoseidonKimchiFp = {
-  name: 'kimchi',
-  test_vectors: [
-    {
-      input: [],
-      output:
-        'a8eb9ee0f30046308abbfa5d20af73c81bbdabc25b459785024d045228bead2f',
-    },
-    {
-      input: [
-        '24868377109704864317484712788444936770816201619055451015684374901992949608178',
-      ],
-      output:
-        'fb5992f65c07f9335995f43fd791d39012ad466717729e61045c297507054f3d',
-    },
-    {
-      input: [
-        '25138500177533925254565157548260087092526215225485178888176592492127995051965',
-        '21606396995955632310354633797836705288048676956201515912792903768825190736997',
-      ],
-      output:
-        'fe2436f2027620a11233318b55d0a117086f09674826d1b7ce08d48ad0736c33',
-    },
-    {
-      input: [
-        '65341190374761678546924136634294878082041006677607592173169799311931861749',
-        '11252918989140053451780076222897667624533869965887021710847899108743182517746',
-        '11993650567890553054152749726648339938264069906532804819530014582284646469491',
-      ],
-      output:
-        '9b1b94444a54af49a7623d1fe1ca72649f0a098daf5704925f024eb6ab0e4b3f',
-    },
-    {
-      input: [
-        '14631744889022021376878179931381842735826405280104225262761584865733804304460',
-        '17766002598385162214718752011395717807114919044757862836313771666836556121863',
-        '20730447567691218079219807511379113661440759319510173788696409574226565989397',
-        '8487103542371264456445106921204847278174307155390594690392717848098873084713',
-      ],
-      output:
-        '47ecd3bf2eed86dcf8d2cef3d7667104689dba4d9bcb54006e0c66f6ec8c5a16',
-    },
-    {
-      input: [
-        '27575266539105063911020305469696466899215781212631453596594551295833393371610',
-        '12607045538084099431197241636641320917344975317615149586616963058448914122136',
-        '4191172324783072531589314287016434391689672887595612443732787485098135060368',
-        '25038504569195635103352006470595733410162902838840568812175073723296969307528',
-        '8326376447499026361851214857371603487559269496692705013743054764013532965513',
-      ],
-      output:
-        '09a2d55277908b7c8214f745b3605f0f9055dcd4c9b594cdd759292c34c3a20c',
-    },
-  ],
-};
