--- src/js_crypto/test_vectors/poseidonLegacy.ts
@@ -1,57 +1,0 @@
-// @gen this file is generated - don't edit it directly
-export { testPoseidonLegacyFp };
-let testPoseidonLegacyFp = {
-  name: 'legacy',
-  test_vectors: [
-    {
-      input: [],
-      output:
-        '1b3251b6912d82edc78bbb0a5c88f0c6fde1781bc3e654123fa6862a4c63e617',
-    },
-    {
-      input: [
-        '24868377109704864317484712788444936770816201619055451015684374901992949608178',
-      ],
-      output:
-        'e99262048e717745fdcd602a3b527e9e7c2e5ba5894158084e2f2e2cb142643d',
-    },
-    {
-      input: [
-        '25138500177533925254565157548260087092526215225485178888176592492127995051965',
-        '21606396995955632310354633797836705288048676956201515912792903768825190736997',
-      ],
-      output:
-        '1f8f9d2fb9547d0254a1c0271ff40042a59927e82fd097d7fa723f8551e8c23a',
-    },
-    {
-      input: [
-        '65341190374761678546924136634294878082041006677607592173169799311931861749',
-        '11252918989140053451780076222897667624533869965887021710847899108743182517746',
-        '11993650567890553054152749726648339938264069906532804819530014582284646469491',
-      ],
-      output:
-        '997810fa8f33879e689c808021d7f1cf302f30f0075754e43dc2f75d76bb3920',
-    },
-    {
-      input: [
-        '14631744889022021376878179931381842735826405280104225262761584865733804304460',
-        '17766002598385162214718752011395717807114919044757862836313771666836556121863',
-        '20730447567691218079219807511379113661440759319510173788696409574226565989397',
-        '8487103542371264456445106921204847278174307155390594690392717848098873084713',
-      ],
-      output:
-        'f930ae5bef2098e3b71949e9878caf5659897f3b9eb1711f296a99cfb740ec3f',
-    },
-    {
-      input: [
-        '27575266539105063911020305469696466899215781212631453596594551295833393371610',
-        '12607045538084099431197241636641320917344975317615149586616963058448914122136',
-        '4191172324783072531589314287016434391689672887595612443732787485098135060368',
-        '25038504569195635103352006470595733410162902838840568812175073723296969307528',
-        '8326376447499026361851214857371603487559269496692705013743054764013532965513',
-      ],
-      output:
-        '461b6ec08fd3a9c37033efd409cf548493b003b2f500db845d5d1fd2ffce1b02',
-    },
-  ],
-};
