+++ src/lib/testing/random.ts
@@ -27,9 +27,9 @@ import { jsLayout } from '../../provable/gen/js-layout.js';
 import { GenericProvable, primitiveTypeMap } from '../../provable/generic.js';
 import { Scalar, PrivateKey, Group } from '../../provable/curve-bigint.js';
 import { Signature } from '../../mina-signer/src/signature.js';
-import { randomBytes } from '../../js_crypto/random.js';
+import { randomBytes } from '../../snarkyjs-bindings/crypto/random.js';
 import { alphabet } from '../../provable/base58.js';
-import { bytesToBigInt } from '../../js_crypto/bigint-helpers.js';
+import { bytesToBigInt } from '../../snarkyjs-bindings/crypto/bigint-helpers.js';
 import { Memo } from '../../mina-signer/src/memo.js';
 import { ProvableExtended } from '../../provable/field-bigint.js';
 import { tokenSymbolLength } from '../../provable/derived-leaves.js';
