+++ src/mina-signer/src/signature.ts
@@ -25,8 +25,8 @@ import {
   withVersionNumber,
 } from '../../provable/binable.js';
 import { base58 } from '../../provable/base58.js';
-import { versionBytes } from '../../js_crypto/constants.js';
-import { Pallas } from '../../js_crypto/elliptic_curve.js';
+import { versionBytes } from '../../snarkyjs-bindings/crypto/constants.js';
+import { Pallas } from '../../snarkyjs-bindings/crypto/elliptic_curve.js';
 
 export {
   sign,
