+++ src/mina-signer/src/transaction-hash.ts
@@ -25,7 +25,7 @@ import { PublicKey, Scalar } from '../../provable/curve-bigint.js';
 import { Signature, SignatureJson } from './signature.js';
 import { blake2b } from 'blakejs';
 import { base58, withBase58 } from '../../provable/base58.js';
-import { versionBytes } from '../../js_crypto/constants.js';
+import { versionBytes } from '../../snarkyjs-bindings/crypto/constants.js';
 
 export {
   hashPayment,
