+++ src/provable/bigint.unit-test.ts
@@ -1,7 +1,10 @@
 import { Field } from './field-bigint.js';
 import { expect } from 'expect';
 import { shutdown } from '../snarky.js';
-import { bytesToBigInt, bigIntToBytes } from '../js_crypto/bigint-helpers.js';
+import {
+  bytesToBigInt,
+  bigIntToBytes,
+} from '../snarkyjs-bindings/crypto/bigint-helpers.js';
 
 function testBigintRoundtrip(x: bigint, size: number) {
   let bytes = bigIntToBytes(x, size);
